refactor(tests): use `define_test_logger` instead of manual creation
And move its definition to a `test` module instead of a `test_utils` to
be in line with the latest naming scheme (only for crates that already
have a `test` module).
