+++ internal/cardano-node/mithril-cardano-node-internal-database/src/test/mod.rs
@@ -10,22 +10,4 @@ pub mod fake_data;
 
 pub use dummy_cardano_db::{DummyCardanoDb, DummyCardanoDbBuilder};
 
-#[cfg(test)]
-pub(crate) struct TestLogger;
-
-#[cfg(test)]
-impl TestLogger {
-    fn from_writer<W: std::io::Write + Send + 'static>(writer: W) -> slog::Logger {
-        use slog::Drain;
-        use std::sync::Arc;
-
-        let decorator = slog_term::PlainDecorator::new(writer);
-        let drain = slog_term::CompactFormat::new(decorator).build().fuse();
-        let drain = slog_async::Async::new(drain).build().fuse();
-        slog::Logger::root(Arc::new(drain), slog::o!())
-    }
-
-    pub(crate) fn stdout() -> slog::Logger {
-        Self::from_writer(slog_term::TestStdoutWriter)
-    }
-}
+mithril_common::define_test_logger!();
