+++ internal/mithril-dmq/src/consumer/pallas.rs
@@ -151,7 +151,7 @@ mod tests {
     };
     use tokio::{net::UnixListener, task::JoinHandle, time::sleep};
 
-    use crate::{test::payload::DmqMessageTestPayload, test_tools::TestLogger};
+    use crate::test::{TestLogger, payload::DmqMessageTestPayload};
 
     use super::*;
