+++ internal/mithril-dmq/src/lib.rs
@@ -9,28 +9,3 @@ pub mod test;
 pub use consumer::{DmqConsumer, DmqConsumerPallas};
 pub use message::DmqMessageBuilder;
 pub use publisher::{DmqPublisher, DmqPublisherPallas};
-
-#[cfg(test)]
-pub(crate) mod test_tools {
-    use std::io;
-    use std::sync::Arc;
-
-    use slog::{Drain, Logger};
-    use slog_async::Async;
-    use slog_term::{CompactFormat, PlainDecorator};
-
-    pub struct TestLogger;
-
-    impl TestLogger {
-        fn from_writer<W: io::Write + Send + 'static>(writer: W) -> Logger {
-            let decorator = PlainDecorator::new(writer);
-            let drain = CompactFormat::new(decorator).build().fuse();
-            let drain = Async::new(drain).build().fuse();
-            Logger::root(Arc::new(drain), slog::o!())
-        }
-
-        pub fn stdout() -> Logger {
-            Self::from_writer(slog_term::TestStdoutWriter)
-        }
-    }
-}
