+++ internal/mithril-metric/src/helper.rs
@@ -106,26 +106,7 @@ macro_rules! build_metrics_service {
 
 #[cfg(test)]
 pub(crate) mod test_tools {
-    use std::{io, sync::Arc};
-
-    use slog::{Drain, Logger};
-    use slog_async::Async;
-    use slog_term::{CompactFormat, PlainDecorator};
-
-    pub struct TestLogger;
-
-    impl TestLogger {
-        fn from_writer<W: io::Write + Send + 'static>(writer: W) -> Logger {
-            let decorator = PlainDecorator::new(writer);
-            let drain = CompactFormat::new(decorator).build().fuse();
-            let drain = Async::new(drain).build().fuse();
-            Logger::root(Arc::new(drain), slog::o!())
-        }
-
-        pub fn stdout() -> Logger {
-            Self::from_writer(slog_term::TestStdoutWriter)
-        }
-    }
+    mithril_common::define_test_logger!();
 }
 
 #[cfg(test)]
