+++ mithril-aggregator/src/artifact_builder/cardano_immutable_files_full.rs
@@ -209,8 +209,7 @@ mod tests {
     use mithril_common::{entities::CompressionAlgorithm, test::double::fake_data};
 
     use crate::{
-        DumbUploader, file_uploaders::MockFileUploader, services::DumbSnapshotter,
-        test_tools::TestLogger,
+        DumbUploader, file_uploaders::MockFileUploader, services::DumbSnapshotter, test::TestLogger,
     };
 
     use super::*;
