+++ mithril-aggregator/src/commands/genesis_command.rs
@@ -368,7 +368,7 @@ mod tests {
 
     use mithril_common::temp_dir;
 
-    use crate::test_tools::TestLogger;
+    use crate::test::TestLogger;
 
     use super::*;
