+++ mithril-aggregator/src/runtime/state_machine.rs
@@ -390,7 +390,7 @@ mod tests {
     use mithril_common::test::double::{Dummy, fake_data};
 
     use crate::entities::OpenMessage;
-    use crate::test_tools::TestLogger;
+    use crate::test::TestLogger;
 
     use super::super::runner::MockAggregatorRunner;
     use super::*;
