+++ mithril-aggregator/src/services/certificate_chain_synchronizer/synchronizer_service.rs
@@ -236,8 +236,8 @@ mod tests {
     use crate::services::{
         MockOpenMessageStorer, MockRemoteCertificateRetriever, MockSynchronizedCertificateStorer,
     };
+    use crate::test::TestLogger;
     use crate::test::mocks::MockCertificateVerifier;
-    use crate::test_tools::TestLogger;
 
     use super::*;
