+++ mithril-aggregator/src/services/snapshotter/ancillary_signer/with_secret_key.rs
@@ -41,7 +41,7 @@ mod tests {
     use std::collections::BTreeMap;
     use std::path::PathBuf;
 
-    use crate::test_tools::TestLogger;
+    use crate::test::TestLogger;
 
     use super::*;
