+++ mithril-aggregator/src/test/mod.rs
@@ -1,3 +1,5 @@
 mod double;
 #[cfg(test)]
 pub mod mocks;
+
+mithril_common::define_test_logger!();
